pyproject.toml
[project]
name = "smart-pandas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.2.3",
    "pydantic>=2.9.2",
    "pyyaml>=6.0.2",
]

[tool.uv]
package = true
dev-dependencies = [
    "pytest>=8.3.3",
    "ruff>=0.7.1",
]

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
